agentic-ecommerce-assistant/ai-agent/requirements.txt
numpy
google-cloud-firestore
google-cloud-aiplatform
google-genai